// src/app/portal/moduls/medical-centers/models/ConfirmCoordination.model.ts
export class ConfirmCoordination {
  IdReference: string;
  Token: string;
  TrxToken: string;
  RDACode: string;
  LocalCode: string;
  ClinicCode: string;
  ClinicName: string;
  ClinicAddress: string;
  SpecialityCode: string;
  SpecialityName: string;
  DocumentType: string;
  Number: string;
  Phone: string;
  Email: string | null;
  YearMonthDay: string;
  Time: string
  ServicesType: string;
  ServicesTypeName: string;
  Price: string;
  Date: string;
  FullNames: string;
  PatientCode: string;
  ProductCode: string;
  Recommendations: string;
  PaymentMethodCode: string;
  PaymentMethodName: string;
  Installments: string;
  UserEmail: string | null;
  PatientName: string;
}